Rename and configure serializer
