+++ persistence-model/src/main/resources/application.conf
@@ -6,6 +6,7 @@ akka {
       institution = "hmda.persistence.serialization.institutions.InstitutionProtobufSerializer"
       filing = "hmda.persistence.serialization.filing.FilingProtobufSerializer"
       submission = "hmda.persistence.serialization.submission.SubmissionProtobufSerializer"
+      larValidated = "hmda.persistence.serialization.validation.LarValidatedProtobufSerializer"
     }
     serialization-bindings {
       "hmda.persistence.messages.events.institutions.InstitutionEvents$InstitutionCreated" = institution
@@ -14,6 +15,7 @@ akka {
       "hmda.persistence.messages.events.institutions.FilingEvents$FilingStatusUpdated" = filing
       "hmda.persistence.messages.events.institutions.SubmissionEvents$SubmissionCreated" = submission
       "hmda.persistence.messages.events.institutions.SubmissionEvents$SubmissionStatusUpdated" = submission
+      "hmda.persistence.messages.events.processing.CommonHmdaValidatorEvents$LarValidated" = larValidated
     }
   }
   persistence {
