+++ persistence-model/src/main/scala/hmda/persistence/serialization/validation/LarValidatedProtobufConverter.scala
@@ -1,4 +1,4 @@
-package hmda.persistence.serialization.filing
+package hmda.persistence.serialization.validation
 
 import hmda.persistence.messages.events.processing.CommonHmdaValidatorEvents.LarValidated
 import hmda.persistence.model.serialization.HmdaFilingEvents.LarValidatedMessage
@@ -7,7 +7,7 @@ import hmda.persistence.model.serialization.SubmissionEvents.SubmissionIdMessage
 import hmda.persistence.serialization.lar.LARProtobufConverter._
 import hmda.persistence.serialization.submission.SubmissionProtobufConverter._
 
-object HmdaFilingProtobufConverter {
+object LarValidatedProtobufConverter {
 
   def larValidatedToProtobuf(obj: LarValidated): LarValidatedMessage = {
     LarValidatedMessage(
