+++ persistence-model/src/test/scala/hmda/persistence/serialization/validation/LarValidatedProtobufSerializerSpec.scala
@@ -1,14 +1,14 @@
-package hmda.persistence.serialization.filing
+package hmda.persistence.serialization.validation
 
 import hmda.model.fi.lar.LarGenerators
-import org.scalatest.{ MustMatchers, PropSpec }
-import org.scalatest.prop.PropertyChecks
 import hmda.model.institution.SubmissionGenerators._
 import hmda.persistence.messages.events.processing.CommonHmdaValidatorEvents.LarValidated
+import org.scalatest.prop.PropertyChecks
+import org.scalatest.{ MustMatchers, PropSpec }
 
-class HmdaFilingProtobufSerializerSpec extends PropSpec with PropertyChecks with MustMatchers with LarGenerators {
+class LarValidatedProtobufSerializerSpec extends PropSpec with PropertyChecks with MustMatchers with LarGenerators {
 
-  val serializer = new HmdaFilingProtobufSerializer()
+  val serializer = new LarValidatedProtobufSerializer()
 
   property("Lar Validated must be serialized to binary and back") {
     forAll(larGen, submissionIdGen) { (lar, submissionId) =>
